streamlined dev and prod build pipelines

+++ src/main.ts
@@ -4,19 +4,37 @@ import { Command } from "@tauri-apps/api/shell";
 import { listen, TauriEvent } from "@tauri-apps/api/event";
 import { socket } from "./stores/socket.store";
 
+if ((import.meta as any).env.PROD) {
+  console.log("Launching server");
+  const cmd = Command.sidecar('binaries/server');
+  cmd.spawn().then((child) => {
+    console.log("Server running");
+    /**
+     * Killing server process when window is closed. Probably won't
+     * work for multi window application
+    */
+    listen(TauriEvent.WINDOW_DESTROYED, () => {
+      socket.disconnect();
+      child.kill();
+      console.log("Server killed");
+    })
 
-const cmd = Command.sidecar('binaries/server');
+    cmd.on("close", () => {
+      console.warn("Server closed")
+    })
 
-cmd.spawn().then((child) => {
-  /**
-   * Killing server process when window is closed. Probably won't
-   * work for multi window application
-   */
-  listen(TauriEvent.WINDOW_DESTROYED, () => {
-    socket.disconnect();
-    child.kill();
+    cmd.on("error", (error) => {
+      console.log(`Server: ${error}`)
+    })
+
+    cmd.stdout.on("data", (line) => {
+      console.log(line);
+    })
+    cmd.stderr.on("data", (line) => {
+      console.error(line);
+    })
   })
-});
+}
 
 
 const app = new App({
